Fix downloads on chrome.
It was flagging the iframes that we used to trigger the downloads as ads. Because of that our downloads would never trigger. We no longer need to use iframes since f-cdn always sets the content-disposition header to trigger the location as a download. We now use the same exact flow of triggering a free game that we do for triggering an owned game.